image url included in blog creation, added simple streamlit app frontend

+++ db/db_blogs.py
@@ -2,13 +2,17 @@ from db.models import DbBlog
 from schemas import BlogPost
 
 from sqlalchemy.orm.session import Session
+from fastapi import HTTPException, status
+from datetime import datetime
 
 
 def create_blog(db: Session, request: BlogPost):
     new_blog = DbBlog(
+        image_url=request.image_url,
         username=request.user_name,
         title=request.title,
-        content=request.content
+        content=request.content,
+        timestamp=datetime.utcnow().replace(microsecond=0)
     )
     db.add(new_blog)
     db.commit()
@@ -22,6 +26,8 @@ def return_all_blogs(db: Session):
 
 def remove_blog(id: int, db: Session):
     blog = db.query(DbBlog).filter(DbBlog.id == id).first()
+    if not blog:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f'id: {id} not found')
     db.delete(blog)
     db.commit()
     return f"deleted blog: {id}, title: {blog.title}"
